requirements.txt
Django
djangorestframework
psycopg2
geoip2
user-agents
django-ratelimit
PyYAML
python-decouple
python-dotenv